+++ test/test_driving_key_field.py
@@ -1,5 +1,6 @@
 """Unit test for DrivingKeyField."""
 import pytest
+
 from diepvries.driving_key_field import DrivingKeyField